Adding ssbM4LArdBits and associated M4L-MaxMSP patches.
The ssbM4LArdBits patch and associated Max4Live device / Max/MSP patch
allows the toggling of specific bits on the ArdCore expander.
Triggering bits will also set the voltage of the DAC output to specific
 amounts. This is a side effect, but could be useful as the amounts are
quite pardicitable.
As the bits may be automated from live, this may be used as a drum
machine trigger for 8 drums or to enable/disable specific effects via
on/off gates.