Variation on 4ths and 5ths patch for the expander
Takes clock in as well as quantized 1v/o CV values in A2. A0 and A1 are
used to determine if the input CV should be shifted up or down and by a
4th or a 5th. The A3 is used to set the frequency of pulses based on
the clock. The A4 and A5 may be used to set the gate width of the gate
pulses (Feed constant voltage or LFO into A4/A5 to use).