Adding ssbSkipper
ssbSkipper: Input Clock or Gate Signals on the clock input.
Depending upon the state of A2/A3, Either output the same gate signal
on D0/D1 or skip it. At the lowest setting it plays all gates. At
higher settings it will skip every other, third, fourth, fifth, sixth
or seventh. At the highest setting it will skip all. If A0/A1 is high
(>50%) the skip will be random, ie it may or may not occur, (50%
chance). The expander A4/A5 may be used to shift the randomness from
50% chance to weighted either toward 10% or 90% (low or high).
The skip none and skip all settings are best with random as they you
get a 50% don't play or a 50% do play for all gates. With the expander
this can then be maniuplated further. Note that the skip and the random
percents may be modulated.